examples/leetcode/problems/28_Find_the_Index_Of_First_Occurence_In_A_String/problem_test.go
package problems

import "testing"

func Test_strStr(t *testing.T) {
	type args struct {
		haystack string
		needle   string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "happy path 1",
			args: args{
				haystack: "sadbutsad",
				needle:   "sad",
			},
			want: 0,
		},
		{
			name: "happy path 2",
			args: args{
				haystack: "leetcode",
				needle:   "leeto",
			},
			want: -1,
		},
		{
			name: "happy path 3",
			args: args{
				haystack: "mississippi",
				needle:   "issipi",
			},
			want: -1,
		},
		{
			name: "happy path 4",
			args: args{
				haystack: "aaa",
				needle:   "aaaa",
			},
			want: -1,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := strStr(tt.args.haystack, tt.args.needle); got != tt.want {
				t.Errorf("strStr() = %v, want %v", got, tt.want)
			}
		})
	}
}
